+++ app/list/[service]/components/ClientComponent.tsx
@@ -9,30 +9,11 @@ import useScroll from '@/hooks/useScroll';
 import Error from '@/utils/ErrorComponent';
 import Navigate from '@/components/Navigate';
 import { useWebtoonList } from '../../hooks/useWebtoosList';
+import { useCategory } from '../../hooks/useCategory';
 
-const day: any = {
-    0: '일요웹툰',
-    1: '월요웹툰',
-    2: '화요웹툰',
-    3: '수요웹툰',
-    4: '목요웹툰',
-    5: '금요웹툰',
-    6: '토요웹툰',
-};
-
-const SelectedCategory = [
-    '전체보기',
-    '월요웹툰',
-    '화요웹툰',
-    '수요웹툰',
-    '목요웹툰',
-    '금요웹툰',
-    '토요웹툰',
-    '일요웹툰',
-];
 
 export default function ClientComponent({ service }: { service: string }) {
-    const [category, setCategory] = useState<string>(day[new Date().getDay()]);
+    const { category, setCategory, SelectedCategory } = useCategory();
 
     const {
         webtoons,
